Refactor base class
